Morse code and NATO alphabet generators.
Should have started this repo earlier, but I didn't.  I do have a
bunch of gists with the work on these generators and some rabbit
holes that didn't quite pan out.

The Morse generator uses WebAudio API to generate blips
corresponding to Morse strings.  The Alphabet generator is a
sprite of the NATO alphabet from Wikipedia, chunked up based on
a modification of the subtitle file that came with it.

The point is to play strings of text as NATO sprites and then
gradually replace that with Morse code, character by character.

Still not sure the Morse code is accurate because I'm making this
to learn Morse code.  If it's not accurate, I'll end up with a
weird Morse accent when I actually get around to keying fluently.

Should still be fairly decipherable, though.  I mean, it's still
a consistent method of keying the characters...